:recycle: rm oss from package path
