--- refarch-gateway/src/main/java/de/muenchen/oss/refarch/gateway/configuration/SecurityProperties.java
@@ -1,16 +1,0 @@
-package de.muenchen.oss.refarch.gateway.configuration;
-
-import java.util.List;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@Getter
-@Setter
-@ConfigurationProperties("refarch.security")
-public class SecurityProperties {
-    /**
-     * List of url patterns excluded from csrf protection.
-     */
-    private List<String> csrfWhitelisted = List.of();
-}
