+++ refarch-gateway/src/main/java/de/muenchen/refarch/gateway/configuration/NoSecurityConfiguration.java
@@ -1,0 +1,23 @@
+package de.muenchen.refarch.gateway.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.web.server.ServerHttpSecurity;
+import org.springframework.security.web.server.SecurityWebFilterChain;
+
+@Configuration
+@Profile("no-security")
+public class NoSecurityConfiguration {
+
+    @Bean
+    public SecurityWebFilterChain springSecurityFilterChain(ServerHttpSecurity http) {
+        return http
+                .authorizeExchange(authorizeExchangeSpec -> authorizeExchangeSpec.anyExchange().permitAll())
+                .cors(corsSpec -> {
+                })
+                .csrf(ServerHttpSecurity.CsrfSpec::disable)
+                .build();
+    }
+
+}
