--- refarch-gateway/src/test/java/de/muenchen/oss/refarch/gateway/TestConstants.java
@@ -1,11 +1,0 @@
-package de.muenchen.oss.refarch.gateway;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class TestConstants {
-
-    public static final String SPRING_TEST_PROFILE = "test";
-
-}
