+++ refarch-gateway/src/test/java/de/muenchen/refarch/gateway/TestConstants.java
@@ -1,0 +1,11 @@
+package de.muenchen.refarch.gateway;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class TestConstants {
+
+    public static final String SPRING_TEST_PROFILE = "test";
+
+}
